+++ azext_prototype/agents/builtin/__init__.py
@@ -2,7 +2,9 @@
 
 from azext_prototype.agents.builtin.advisor import AdvisorAgent
 from azext_prototype.agents.builtin.app_developer import AppDeveloperAgent
-from azext_prototype.agents.builtin.application_architect import ApplicationArchitectAgent
+from azext_prototype.agents.builtin.application_architect import (
+    ApplicationArchitectAgent,
+)
 from azext_prototype.agents.builtin.bicep_agent import BicepAgent
 from azext_prototype.agents.builtin.biz_analyst import BizAnalystAgent
 from azext_prototype.agents.builtin.cloud_architect import CloudArchitectAgent
@@ -11,7 +13,9 @@ from azext_prototype.agents.builtin.csharp_developer import CSharpDeveloperAgent
 from azext_prototype.agents.builtin.data_architect import DataArchitectAgent
 from azext_prototype.agents.builtin.doc_agent import DocumentationAgent
 from azext_prototype.agents.builtin.governor_agent import GovernorAgent
-from azext_prototype.agents.builtin.infrastructure_architect import InfrastructureArchitectAgent
+from azext_prototype.agents.builtin.infrastructure_architect import (
+    InfrastructureArchitectAgent,
+)
 from azext_prototype.agents.builtin.monitoring_agent import MonitoringAgent
 from azext_prototype.agents.builtin.project_manager import ProjectManagerAgent
 from azext_prototype.agents.builtin.python_developer import PythonDeveloperAgent